Rename to `mpasa` in component configuration
`cime/CIME/Tools/Makefile` explicitly queries the value of `CAM_DYCORE`.
If the value is `mpas`, it attempts to build MPAS by using old rules and fails.

There are some ways to work around this without modifying
`cime/CIME/Tools/Makefile` and breaking backward compatibility for CAM. One of
them is to rename the value to `mpasa`.