Set the git version string of MPAS to a more descriptive value
Before, it would always show "N/A" in the log, which is not very helpful.

Now, if the git submodule of MPAS is aligned at a tag, it will just show the tag
name (e.g., "v8.2.1"). Otherwise, if the git submodule of MPAS is not aligned at
a tag, it will show the nearest tag name, the number of commits after that tag
and the abbreviated commit hash (e.g., "v8.0.1-54-ged6f8e39e").