Make PIO version selection more robust
The `USE_PIO2` macro is an implementation detail of MPAS. We should not
let CIME to set it for us, and CIME really should not have included
that logic in the first place.

When CIME invokes its Makefile, `PIO_VERSION` is always supplied as a
command line argument. Therefore, we can reliably use it to determine
if the `USE_PIO2` macro should be set.