Fix linter errors about trailing backslashes
Original linter error message was:
C051 Trailing backslash

The sequence of a trailing backslash followed by a newline will be interpreted as
a continued line by the C preprocessor (`cpp`). According to its documentation,
there is no way to prevent a backslash at the end of a line from being interpreted as
a backslash-newline sequence.

Unfortunately, some code comments contain LaTeX markups to express equations.
In LaTeX, `\\` is used to indicate the end of a row, but it is being misinterpreted
by the C preprocessor as a continued line.

Work around this problem by adding the optional length argument (e.g., `[0pt]`)
to `\\`.